mblog/requirements.txt
asgiref==3.2.10
Django==2.1.10
django-markdown-deux==1.0.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Jinja2==2.11.2
markdown2==2.3.9
MarkupSafe==1.1.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
